Rearrange left nav groups
