+++ packages/front-end/package.json
@@ -48,7 +48,7 @@
     "react-dom": "^17.0.1",
     "react-dropzone": "^11.3.2",
     "react-hook-form": "^7.15.0",
-    "react-icons": "^4.2.0",
+    "react-icons": "^4.3.1",
     "react-paginate": "^8.0.0",
     "react-select": "^5.2.1",
     "react-syntax-highlighter": "^15.4.3",
